* Fewer differences to upstream

* Delete publish_crates.yml

* Remove extra parameters

* Better memory using!

* Converting to same format may continue

Now can set to using compress when convertor confused.

+++ src/error.rs
@@ -5,7 +5,7 @@ pub type Result<T> = std::result::Result<T, CaesiumError>;
 #[derive(Debug, Clone)]
 pub struct CaesiumError {
     pub message: String,
-    pub code: u64,
+    pub code: u32,
 }
 
 impl fmt::Display for CaesiumError {
